refactor(stacks): extract StacksOverview from route file into components/
Route files should only declare the route. Moving StacksOverview to
src/components/stacks/ puts the component and its tests where the
router can't mistake them for route definitions.
